Getting testing infrastructure set up.
Elm has a requirement where module names match file names.  This conflicts
with exercism's requirement for example files to be named example.

This is why we included ruby for our _test/bootstrap.rb.
It strips example files of 'example' in their filename, so exercism is happy, and so is Elm.

Also includes RNATranscription Test.